refactor weekly summary route
