+++ apps/web/app/api/cron/weekly-summary/lib/product.ts
@@ -1,0 +1,93 @@
+import { prisma } from "@formbricks/database";
+import { TWeeklySummaryProductData } from "@formbricks/types/weeklySummary";
+
+export const getProductsByTeamId = async (teamId: string): Promise<TWeeklySummaryProductData[]> => {
+  const sevenDaysAgo = new Date();
+  sevenDaysAgo.setDate(sevenDaysAgo.getDate() - 7);
+
+  return await prisma.product.findMany({
+    where: {
+      teamId: teamId,
+    },
+    select: {
+      id: true,
+      name: true,
+      environments: {
+        where: {
+          type: "production",
+        },
+        select: {
+          id: true,
+          surveys: {
+            where: {
+              NOT: {
+                AND: [
+                  { status: "completed" },
+                  {
+                    responses: {
+                      none: {
+                        createdAt: {
+                          gte: sevenDaysAgo,
+                        },
+                      },
+                    },
+                  },
+                ],
+              },
+              status: {
+                not: "draft",
+              },
+            },
+            select: {
+              id: true,
+              name: true,
+              questions: true,
+              status: true,
+              responses: {
+                where: {
+                  createdAt: {
+                    gte: sevenDaysAgo,
+                  },
+                },
+                select: {
+                  id: true,
+                  createdAt: true,
+                  updatedAt: true,
+                  finished: true,
+                  data: true,
+                },
+                orderBy: {
+                  createdAt: "desc",
+                },
+              },
+              displays: {
+                where: {
+                  createdAt: {
+                    gte: sevenDaysAgo,
+                  },
+                },
+                select: {
+                  id: true,
+                },
+              },
+            },
+          },
+        },
+      },
+      team: {
+        select: {
+          memberships: {
+            select: {
+              user: {
+                select: {
+                  email: true,
+                  notificationSettings: true,
+                },
+              },
+            },
+          },
+        },
+      },
+    },
+  });
+};
