+++ apps/web/app/api/cron/weekly-summary/lib/team.ts
@@ -1,0 +1,10 @@
+import { prisma } from "@formbricks/database";
+
+export const getTeamIds = async (): Promise<string[]> => {
+  const teams = await prisma.team.findMany({
+    select: {
+      id: true,
+    },
+  });
+  return teams.map((team) => team.id);
+};
